修复代码中的一些错误，包括拼写错误和逻辑错误。

+++ duilib/Utils/FileUtil.cpp
@@ -23,7 +23,7 @@ bool FileUtil::ReadFileData(const FilePath& filePath, std::vector<uint8_t>& file
     if (f != nullptr) {
         isReadOk = true;
         ::fseek(f, 0, SEEK_END);
-        int fileSize = ::ftell(f);
+        long fileSize = ::ftell(f);
         ::fseek(f, 0, SEEK_SET);
         if (fileSize > 0) {
             fileData.resize((size_t)fileSize);
